requirements.txt
supabase==2.0.3
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
